remove unnecessary statements and move test case to broker.ltw package
